move functions to get household count data from `property_density.py` to `household_count.py` - add new function to `household_count.py` to load and transform household count for LSOAs and DataZones
